+++ src/faucet/views/components/mod.rs
@@ -4,4 +4,5 @@ pub mod faucet_description;
 pub mod icons;
 pub mod layout;
 pub mod nav;
+pub mod rpc_selector;
 pub mod transaction;